Fixed: based on code review
